requirements.txt
apache-airflow-providers-papermill==3.0.0
ipykernel==6.15.1
notebook==6.4.12
psycopg2-binary==2.9.3
SQLAlchemy==1.4.40
python-dotenv==0.20.0
pandas==1.4.3
numpy==1.23.2
matplotlib==3.5.3
seaborn==0.11.2
nltk==3.7
scikit-learn==1.1.2
wordcloud==1.8.2.2